+++ src/dispatcher.ts
@@ -1,5 +1,5 @@
-import { CommandOptions } from '../types/commandOptions';
-import { CodeceptionCommand } from './codeception';
+import { CommandOptions } from "./interfaces";
+import { CodeceptionCommand } from "./commands";
 
 export default function Dispatcher (commandOptions: CommandOptions = {}) {
     return new CodeceptionCommand(commandOptions);