+++ src/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './node-scopes.helper';
+export * from './symbol-node.helper';