+++ src/interfaces/command-options.interface.ts
@@ -1,0 +1,9 @@
+
+/**
+ * Interfaces
+ */
+export interface CommandOptions {
+    runAll?: boolean;
+    runFile?: boolean;
+    runMethod?: boolean;
+}