+++ src/types/commandOptions.ts
@@ -1,8 +1,0 @@
-// Available options for command
-export interface CommandOptions {
-    runAll?: boolean;
-    runFile?: boolean;
-    runClass?: boolean;
-    method?: string;
-    uri?: string;
-}