+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## [1.1.0] - 2019-05-12
+
+### Added
+- Added support for running a specific method.
+    - Place your cursor within a function and press `cmd+k + cmd+e`
+
+### Fixed
+- Fixed issue with command suffix not being appended to the `run-all` command
+
+[1.1.0]: https://github.com/joelwmale/vscode-codeception/compare/1.0,0...1.1.0
+
 ## [1.0.0] - 2019-05-10
 
 - Initial release